ai_segmentation/code2/requirements.txt
# PyTorch UNet CHAOS-CT 项目依赖

# 深度学习框架
torch>=1.7.1
torchvision>=0.8.2

# 图像处理
opencv-python>=4.5.0
Pillow>=8.0.0

# 数据处理
numpy>=1.19.0
pandas>=1.2.0

# 医学图像处理
SimpleITK>=2.0.0

# 进度和可视化
tqdm>=4.60.0
matplotlib>=3.3.0

# 科学计算
scipy>=1.6.0
scikit-image>=0.18.0

# 可选：Vitis AI 量化支持
# vai_q_pytorch  # 需要Vitis AI环境
# xir  # 需要Vitis AI环境

# 开发工具（可选）
# tensorboardX>=2.1
# thop>=0.1.0  # FLOPs计算